First step to transition to 0.66pre. Not working yet.
