+++ Scripts/Mira/Subcontrollers/Squads/SquadStates/MiraSquadAttackGatheringUpState.ts
@@ -1,0 +1,9 @@
+import { MiraSquadAttackState } from "./MiraSquadAttackState";
+import { MiraSquadGatheringUpState } from "./MiraSquadGatheringUpState";
+
+export class MiraSquadAttackGatheringUpState extends MiraSquadGatheringUpState {
+    protected onGatheredUp(): void {
+        this.squad.AttackTargetCell = this.squad.CurrentTargetCell;
+        this.squad.SetState(new MiraSquadAttackState(this.squad));
+    }
+}
