--- Scripts/Mira/Utils/00_Common.ts
@@ -1,6 +1,0 @@
-
-// abstract class FsmState {
-//     abstract OnEntry(): void;
-//     abstract OnExit(): void;
-//     abstract Tick(tickNumber: number): void;
-// }
